+++ src/vec3_scale.c
@@ -6,13 +6,13 @@
 /*   By: scollon <scollon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/08 12:58:18 by scollon           #+#    #+#             */
-/*   Updated: 2016/02/08 13:18:56 by scollon          ###   ########.fr       */
+/*   Updated: 2016/02/09 16:30:58 by scollon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libvect.h"
 
-t_vect3d	vec3_scale(t_vect3d v3, float scale)
+t_vect3d	vec3_scale(t_vect3d v3, double scale)
 {
 	t_vect3d	r;
